refactor!: convert ALL 29 infrastructure adapters to TypeScript — src/ is 100% TS
29 infrastructure adapter .js files converted to .ts:
- adapterValidation, CasBlobAdapter, CasSeekCacheAdapter
- CborCheckpointStoreAdapter, CborPatchJournalAdapter, CborIndexStoreAdapter
- ClockAdapter, ConsoleLogger, NoOpLogger
- NoOpEffectSink, ConsoleEffectSink, ChunkEffectSink
- LoggerObservabilityBridge, lazyCasInit, requireCapabilities
- BunHttpAdapter, DenoHttpAdapter, NodeHttpAdapter, httpAdapterUtils
- NodeCryptoAdapter, WebCryptoAdapter, TrustCryptoAdapter, sha1sync
- CborDecodeTransform, CborEncodeTransform, GitBlobWriteTransform
- IndexShardEncodeTransform, TreeAssemblerSink, CborCodec

src/ .js: 0. src/ .ts: 374. THE SOURCE DIRECTORY IS 100% TYPESCRIPT.

All import paths updated across src/, bin/, test/, index.js, browser.js.
ESLint globals updated for Bun/Deno runtime types.
Duplicate imports merged. Stale eslint-disable comments removed.

Source tsc: 0. Test tsc: 0. Lint: CLEAN. Tests: 356/356.

BREAKING CHANGE: All infrastructure adapter imports must use .ts extension.

+++ src/infrastructure/adapters/httpAdapterUtils.ts
@@ -41,7 +41,7 @@ export async function readStreamBody(bodyStream: ReadableStream<Uint8Array>): Pr
     if (result.done) {
       break;
     }
-    const chunk = result.value as Uint8Array;
+    const chunk = result.value;
     total += chunk.byteLength;
     if (total > MAX_BODY_BYTES) {
       await reader.cancel();
